requirements.txt
Flask==2.0.3
Flask-MySQLdb==1.0.0
bcrypt